move AtdfRecord to stdf_types
