+++ src/lib.rs
@@ -42,8 +42,8 @@ pub mod atdf_file;
 #[cfg(test)]
 mod tests {
     use crate::*;
-    use atdf_file::atdf_record_field::*;
     use stdf_record_type::*;
+    use stdf_types::atdf_record_field::*;
     use stdf_types::ByteOrder;
 
     #[test]
@@ -1154,7 +1154,7 @@ mod tests {
 
     #[test]
     fn test_atdf_field_req_count() {
-        assert_eq!(3, atdf_file::count_reqired(&PTR_FIELD));
-        assert_eq!(0, atdf_file::count_reqired(&GDR_FIELD));
+        assert_eq!(3, stdf_types::count_reqired(&PTR_FIELD));
+        assert_eq!(0, stdf_types::count_reqired(&GDR_FIELD));
     }
 }
